+++ Sources/Controllers/ALKConversationViewController.swift
@@ -91,6 +91,10 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
     fileprivate var profanityFilter: ProfanityFilter?
     var activeTextField: UITextField?
     var documentManager = ALKDocumentManager()
+    lazy var photoPicker = ALKPhotoPicker(
+        localizationFileName: localizedStringFileName,
+        selectionLimit: configuration.chatBar.photosSelectionLimit
+    )
 
     fileprivate enum ActionType: String {
         case link
@@ -706,6 +710,7 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
         chatBar.accessibilityIdentifier = "chatBar"
         chatBar.setComingSoonDelegate(delegate: view)
         documentManager.delegate = self
+        photoPicker.delegate = self
         chatBar.action = { [weak self] action in
 
             guard let weakSelf = self else {
@@ -830,11 +835,15 @@ open class ALKConversationViewController: ALKBaseViewController, Localizable {
                     ALUtilityClass.showAlertMessage(msg, andTitle: title)
                 }
             case .showImagePicker:
-                guard let vc = ALKCustomPickerViewController.makeInstanceWith(delegate: weakSelf, and: weakSelf.configuration)
-                else {
-                    return
+                if #available(iOS 14, *), weakSelf.configuration.isNewSystemPhotosUIEnabled {
+                    weakSelf.photoPicker.openGallery(from: weakSelf)
+                } else {
+                    guard let vc = ALKCustomPickerViewController.makeInstanceWith(delegate: weakSelf, and: weakSelf.configuration)
+                    else {
+                        return
+                    }
+                    weakSelf.present(vc, animated: false, completion: nil)
                 }
-                weakSelf.present(vc, animated: false, completion: nil)
             case .showLocation:
                 let storyboard = UIStoryboard.name(storyboard: UIStoryboard.Storyboard.mapView, bundle: Bundle.applozic)